refactor(dipu): resolve comments from PR #771
